Refactored routes implementations
