+++ src/components/auth/PrivateRoute.tsx
@@ -1,18 +1,24 @@
 import React from 'react';
 import {useSelector} from 'react-redux';
-import {Route, Redirect, RouteProps} from 'react-router-dom';
-import {RootState} from '../../store';
+import {Redirect, Route, RouteProps} from 'react-router-dom';
 
 import AppPaths from '../../const/paths';
+import {RootState} from '../../state/store';
+import {RouteComponentProps} from './props';
 
-interface Props extends RouteProps {
-    component: any;
-}
+type PrivateRouteProps = RouteProps & RouteComponentProps;
 
-export const PrivateRoute = ({component: Component, ...rest}:Props) => {
-  const {authenticated} = useSelector((state: RootState) => state.auth);
+export const PrivateRoute = ({component: Component, ...rest}: PrivateRouteProps) => {
+  const {user} = useSelector((state: RootState) => state.auth);
+
+  if (!user) {
+    return <Redirect to={AppPaths.SIGN_IN}/>;
+  }
 
   return (
-    <Route {...rest} render={(props) => authenticated ? <Component {...props} /> : <Redirect to={AppPaths.SIGN_IN}/>} />
+    <Route
+      render={(props) => <Component {...props} />}
+      {...rest}
+    />
   );
 };
