+++ src/components/auth/PublicRoute.tsx
@@ -1,19 +1,25 @@
 import React from 'react';
 import {useSelector} from 'react-redux';
-import {Route, Redirect, RouteProps} from 'react-router-dom';
-import {RootState} from '../../store';
-
+import {Redirect, Route, RouteProps} from 'react-router-dom';
 import AppPaths from '../../const/paths';
 
-interface Props extends RouteProps {
-  component: any;
-}
+import {RootState} from '../../state/store';
+import {RouteComponentProps} from './props';
 
-export const PublicRoute = ({component: Component, ...rest}:Props) => {
-  const {authenticated} = useSelector((state: RootState) => state.auth);
+type PublicRouteProps = RouteProps & RouteComponentProps;
+
+export const PublicRoute = ({component: Component, ...rest}: PublicRouteProps) => {
+  const {user} = useSelector((state: RootState) => state.auth);
+
+  if (user) {
+    // FIXME: Testing placeholder, to be removed
+    return <Redirect to={AppPaths.CALC}/>;
+  }
 
   return (
-    <Route {...rest} render={(props) => !authenticated ? <Component {...props} /> :
-      <Redirect to={AppPaths.CALC} />} />
+    <Route
+      render={(props) => <Component {...props} />}
+      {...rest}
+    />
   );
 };
