+++ src/components/auth/props.ts
@@ -1,0 +1,6 @@
+import React from 'react';
+
+
+export type RouteComponentProps = {
+  component: React.FunctionComponent,
+}
